requirements.txt
azure-functions
azure-storage-blob
matplotlib
pandas
requests